requirements.txt
rich>=13.7.0         
Pillow>=10.0.0       
exifread>=3.0.0       
# binwalk>=2.3.4       # Uncomment if you want the Python API (requires system binwalk)
